automatically backed up by learn

+++ index.js
@@ -19,5 +19,5 @@ function dec(n) {
   return n;
 }
 function makeInt(n) {
-  return parseFloat(n, 10);
+  return parseInt(n, 10);
 }
